chore[litemall-vue]: 简化登录组件，同时添加测试账号提醒信息
