--- litemall-vue/src/views/login/login-footer.vue
@@ -1,18 +1,0 @@
-<template>
-  <div class="text-desc text-center bottom_positon">技术支持: litemall</div>
-</template>
-
-<script>
-export default {
-  name: 'login-footer'
-};
-</script>
-
-
-<style scoped>
-.bottom_positon {
-  position: absolute;
-  bottom: 30px;
-  width: 100%;
-}
-</style>
